Migrate REPL logic into separate file
